// src/shared/interfaces/florist.interface.ts
export interface Florist {
    id: number;
    status: string;
    ref: string;
    florist_name: string;
    contact_number: string;
    address: string;
    city: string;
    postcode: string;
    call_outcome: string;
    social_media_link: string;
    product_type: string;
    product_price: string;
    delivery_fee: string;
    sell_extras: string;
    popularity_trend: string;
    preferred_communication: string;
    has_website: string;
    member_of_other_networks: string;
    flower_supplier: string;
    interested_in_free_website: string;
    discount_offer: null;
    florist_rep: string;
    additional_info: string | null;
    floristcode: string;
    photo: string | null;
    description: string | null;
    collections: string | null;
    meta_description: string | null;
    page_title: string | null;
    photo_url: string | null;
    province: string;
    imported: number;
    shopifyid: string;
    listing_status: boolean | null;
};